+++ SKPhotoBrowser/SKAnimator.swift
@@ -102,10 +102,10 @@ class SKAnimator: NSObject, SKPhotoBrowserAnimatorDelegate {
             let offsetY = scrollView.center.y - (scrollView.bounds.height/2)
             let frame = CGRect(
                 x: scrollFrame.origin.x - contentOffset.x,
-                y: scrollFrame.origin.y + contentOffset.y + offsetY,
+                y: scrollFrame.origin.y + contentOffset.y + offsetY - scrollView.contentOffset.y,
                 width: scrollFrame.width,
                 height: scrollFrame.height)
-            
+
             resizableImageView.image = image.rotateImageByOrientation()
             resizableImageView.frame = frame
             resizableImageView.alpha = 1.0
@@ -138,6 +138,11 @@ private extension SKAnimator {
             let height = width / imageRatio
             let yOffset = (SKMesurement.screenHeight - height) / 2
             return CGRect(x: 0, y: yOffset, width: width, height: height)
+            
+        } else if SKPhotoBrowserOptions.longPhotoWidthMatchScreen && imageRatio <= 1.0 {
+            let height = SKMesurement.screenWidth / imageRatio
+            return CGRect(x: 0.0, y: 0, width: SKMesurement.screenWidth, height: height)
+            
         } else {
             let height = SKMesurement.screenHeight
             let width = height * imageRatio