+++ SKPhotoBrowser/SKZoomingScrollView.swift
@@ -119,14 +119,17 @@ open class SKZoomingScrollView: UIScrollView {
         
         let xScale = boundsSize.width / imageSize.width
         let yScale = boundsSize.height / imageSize.height
-        let minScale: CGFloat = min(xScale, yScale)
+        var minScale: CGFloat = min(xScale, yScale)
         var maxScale: CGFloat = 1.0
         
         let scale = max(UIScreen.main.scale, 2.0)
         let deviceScreenWidth = UIScreen.main.bounds.width * scale // width in pixels. scale needs to remove if to use the old algorithm
         let deviceScreenHeight = UIScreen.main.bounds.height * scale // height in pixels. scale needs to remove if to use the old algorithm
         
-        if imageView.frame.width < deviceScreenWidth {
+        if SKPhotoBrowserOptions.longPhotoWidthMatchScreen && imageView.frame.height >= imageView.frame.width {
+            minScale = 1.0
+            maxScale = 2.5
+        } else if imageView.frame.width < deviceScreenWidth {
             // I think that we should to get coefficient between device screen width and image width and assign it to maxScale. I made two mode that we will get the same result for different device orientations.
             if UIApplication.shared.statusBarOrientation.isPortrait {
                 maxScale = deviceScreenHeight / imageView.frame.width
@@ -155,7 +158,7 @@ open class SKZoomingScrollView: UIScrollView {
         */
         
         // reset position
-        imageView.frame = CGRect(x: 0, y: 0, width: imageView.frame.size.width, height: imageView.frame.size.height)
+        imageView.frame.origin = CGPoint.zero
         setNeedsLayout()
     }
     
@@ -191,14 +194,23 @@ open class SKZoomingScrollView: UIScrollView {
 
             var imageViewFrame: CGRect = .zero
             imageViewFrame.origin = .zero
-            imageViewFrame.size = image.size
-
+            // long photo
+            if SKPhotoBrowserOptions.longPhotoWidthMatchScreen && image.size.height >= image.size.width {
+                let imageHeight = SKMesurement.screenWidth / image.size.width * image.size.height
+                imageViewFrame.size = CGSize(width: SKMesurement.screenWidth, height: imageHeight)
+            } else {
+                imageViewFrame.size = image.size
+            }
             imageView.frame = imageViewFrame
 
             contentSize = imageViewFrame.size
-            
             setMaxMinZoomScalesForCurrentBounds()
+            
+        } else {
+            // change contentSize will reset contentOffset, so only set the contentsize zero when the image is nil
+            contentSize = CGSize.zero
         }
+    
         setNeedsLayout()
     }